requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python-headless
pytesseract
python-multipart
pdf2image
pydantic
scikit-learn
